requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
colorama==0.4.6
comtypes==1.4.13
contourpy==1.3.0
cycler==0.12.1
exceptiongroup==1.3.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
llvmlite==0.43.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mediapipe==0.10.21
ml_dtypes==0.5.3
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
ollama==0.6.0
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==4.25.8
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.12.3
pydantic_core==2.41.4
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyparsing==3.2.3
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
regex==2025.10.23
requests==2.32.5
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
zipp==3.23.0
